codegen: move `HasCtxMethods` into top module
