+++ compiler/hash-codegen/src/lower/operands.rs
@@ -10,8 +10,8 @@ use super::{locals::LocalRef, place::PlaceRef, utils, FnBuilder};
 use crate::{
     common::MemFlags,
     traits::{
-        builder::BlockBuilderMethods, constants::ConstValueBuilderMethods, ctx::HasCtxMethods,
-        layout::LayoutMethods, ty::TypeBuilderMethods, CodeGenObject, Codegen,
+        builder::BlockBuilderMethods, constants::ConstValueBuilderMethods, layout::LayoutMethods,
+        ty::TypeBuilderMethods, CodeGenObject, Codegen, HasCtxMethods,
     },
 };
